update country to region
