+++ etc/config.xml
@@ -17,7 +17,7 @@
                 <active>0</active>
                 <debug>1</debug>
                 <model>OnePicaAffirmGatewayFacade</model>
-                <country>United States</country>
+                <region>United States</region>
                 <mode>sandbox</mode>
                 <private_api_key_sandbox backend_model="Magento\Config\Model\Config\Backend\Encrypted" />
                 <private_api_key_production backend_model="Magento\Config\Model\Config\Backend\Encrypted" />
@@ -29,7 +29,7 @@
                 <partial_capture>0</partial_capture>
                 <disable_for_backordered_items>0</disable_for_backordered_items>
                 <allowspecific>1</allowspecific>
-                <specificcountry>US,AS,GU,MP,PR,VI</specificcountry>
+                <specificcountry>US,AS,GU,MP,PR,VI,CA</specificcountry>
                 <allowspecificcurrency>1</allowspecificcurrency>
                 <currency>USD</currency>
                 <can_authorize>1</can_authorize>
